feat(ci): Ajout intégration continue GitHub Actions dans branche dev (#3)
I had forgotten to clarify that PR #2 was a stacked PR, proposed to be merged into the dev-finalisation branch (the one of PR #1) instead of dev. Therefore, in order to merge both to dev, PR #2 should have been merged before merging PR #1.

That's ok, we can still merge it to the dev branch, now. (I set the base branch to dev)